// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Category = {
  STREET_ART: 'STREET_ART',
  BUILDING: 'BUILDING',
  RUIN: 'RUIN'
} as const

export type Category = (typeof Category)[keyof typeof Category]


export const Status = {
  ACTIVE: 'ACTIVE',
  DESTROYED: 'DESTROYED'
} as const

export type Status = (typeof Status)[keyof typeof Status]